Move useMerkle into src/ui
