+++ apps/council-frontend/src/ui/airdrop/ReviewClaim/ReviewTransaction.tsx
@@ -5,7 +5,7 @@ import React, { ReactElement, useCallback, useRef, useState } from "react";
 import toast from "react-hot-toast";
 import { isValidAddress } from "src/base/isValidAddress";
 import { ETHERSCAN_TRANSACTION_DOMAIN } from "src/integrations/etherscan";
-import { MerkleRewardType, useMerkleInfo } from "src/merkle/useMerkleInfo";
+import { MerkleRewardType, useMerkleInfo } from "src/ui/merkle/useMerkleInfo";
 import { AirdropAmountCard } from "src/ui/airdrop/AirdropAmountCard/AirdropAmountCard";
 import { StepCard } from "src/ui/airdrop/StepCard/StepCard";
 import { useClaimAndDepositAirdrop } from "src/ui/airdrop/useClaimAndDepositAirdrop";
