+++ apps/council-frontend/src/ui/voting/useVote.ts
@@ -10,7 +10,7 @@ import { ethers, Signer } from "ethers";
 import { addressesJson } from "src/elf-council-addresses";
 import { coreVotingContract } from "src/contracts";
 import { MerkleProof } from "src/merkle/MerkleProof";
-import { MerkleRewardType, useMerkleInfo } from "src/merkle/useMerkleInfo";
+import { MerkleRewardType, useMerkleInfo } from "src/ui/merkle/useMerkleInfo";
 import { useLatestBlockNumber } from "src/ui/ethereum/useLatestBlockNumber";
 import { Ballot } from "src/ui/voting/Ballot";
 import { useLockingVaultVotingPower } from "src/ui/voting/useLockingVaultVotingPower";
